+++ lib/xcodeproj/project/object/group.rb
@@ -300,19 +300,6 @@ module Xcodeproj
           file
         end
 
-        # Creates a new group to represent a `xcdatamodel` file.
-        #
-        # @return [XCVersionGroup] The new group.
-        #
-        def new_xcdatamodel_group(xcdatamodel_path)
-          g = @project.new(XCVersionGroup)
-          g.path = xcdatamodel_path
-          g.version_group_type = 'wrapper.xcdatamodel'
-          file = g.new_file(xcdatamodel_path.sub(/xcdatamodeld$/, 'xcdatamodel'))
-          g.current_version = file
-          g
-        end
-
         # Traverses the children groups and finds the group with the given
         # path, if exists.
         #